PARSER_NEW/requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
astor==0.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
numpy==2.0.1
packaging==24.1
page==0.2
pandas==2.2.2
pillow==10.4.0
playwright==1.45.1
pluggy==1.5.0
pretty==0.1
pyee==11.1.0
Pygments==2.18.0
pyodbc==5.1.0
pyparsing==3.1.2
pytest==8.3.2
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.5.1
pytest-pretty-terminal==1.1.0
pytest-pspec==0.0.4
pytest-rerunfailures==14.0
pytest-sugar==1.0.0
pytest-testdox==3.1.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
requests==2.32.3
rerun==1.0.30
rich==13.7.1
screeninfo==0.8.1
six==1.16.0
termcolor==2.4.0
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
